+++ packages/SequenceTranslator/src/polytool/conversion/rule-manager.ts
@@ -6,6 +6,7 @@ import {_package, applyNotationProviderForCyclized} from '../../package';
 import {getHelmHelper} from '@datagrok-libraries/bio/src/helm/helm-helper';
 import {doPolyToolConvert} from './pt-conversion';
 import {NOTATION} from '@datagrok-libraries/bio/src/utils/macromolecule/consts';
+import {RuleCards} from './pt-rule-cards';
 
 
 const TAB_LINKS = 'Links';
@@ -22,12 +23,15 @@ export class RulesManager {
   homoDimerInput: DG.InputBase;
   heteroDimerInput: DG.InputBase;
 
+  linkCards: RuleCards[];
+
   // every rule set will have its editor instance
   private static instances: Record<string, RulesManager> = {};
 
   protected constructor(rules: Rules, fileName: string) {
     this.rules = rules;
     this.linkRuleDataFrame = this.rules.getLinkRulesDf();
+
     this.synthRuleDataFrame = this.rules.getSynthesisRulesDf();
     this.fileName = fileName;
 
@@ -96,8 +100,9 @@ export class RulesManager {
     grok.shell.info(`Polytool rules at ${this.fileName} was updated`);
   }
 
-  private createGridDiv(name: string, grid: DG.Grid) {
+  private createGridDiv(name: string, grid: DG.Grid, tooltipMsg?: string) {
     const header = ui.h1(name, 'polytool-grid-header');
+    ui.tooltip.bind(header, tooltipMsg);
     header.style.marginTop = '10px';
     header.style.marginRight = '10px';
     grid.root.style.height = '100%';
@@ -177,11 +182,45 @@ export class RulesManager {
     inputsTabControl: DG.TabControl;
 
     const linksGridDiv = this.createGridDiv('Rules',
-      this.linkRuleDataFrame.plot.grid({showAddNewRowIcon: true}));
-    const linkExamples = this.createGridDiv('Examples', await this.getLinkExamplesGrid());
+      this.linkRuleDataFrame.plot.grid({showAddNewRowIcon: true}),
+      'specification for monomers to link and linking positions');
+    const linkExamples = this.createGridDiv('Examples', await this.getLinkExamplesGrid(),
+      'specification for monomers to link and linking positions');
     linksGridDiv.style.width = '50%';
     linkExamples.style.width = '50%';
-    const links = ui.divH([linksGridDiv, linkExamples]);
+    const header = ui.h1('Monomers', 'polytool-grid-header');
+    ui.tooltip.bind(header, 'Click different cobination to see how monomers will link');
+    this.linkCards = await this.rules.getLinkCards();
+    const gridDiv: HTMLElement = ui.splitV([
+      ui.box(
+        header,
+        {style: {maxHeight: '30px'}},
+      ),
+      this.linkCards[0].root,
+    ]);
+
+    this.linkCards[0].render();
+    await this.linkCards[0].reset();
+
+    this.linkRuleDataFrame.currentRowIdx = 0;
+    this.linkRuleDataFrame.onCurrentRowChanged.subscribe(async () => {
+      const idx = this.linkRuleDataFrame.currentRowIdx;
+      if (idx !== -1) {
+        ui.empty(gridDiv);
+        gridDiv.append(ui.splitV([
+          ui.box(
+            header,
+            {style: {maxHeight: '30px'}},
+          ),
+          this.linkCards[idx].root,
+        ]));
+      }
+
+      this.linkCards[idx].render();
+      await this.linkCards[idx].reset();
+    });
+
+    const links = ui.splitH([linksGridDiv, gridDiv], null, true);
 
     const reactionsGridDiv = this.createGridDiv('Rules',
       this.synthRuleDataFrame.plot.grid({showAddNewRowIcon: true}));
@@ -200,6 +239,13 @@ export class RulesManager {
       'Reactions': reactions,
       'Dimers': dimerInputsDiv,
     }, false);
+
+    ui.tooltip.bind(inputsTabControl.getPane(TAB_LINKS).header,
+      'Specify rules to link monomers based on HELM notation');
+    ui.tooltip.bind(inputsTabControl.getPane(TAB_REACTIONS).header,
+      'Specify rules to perform reactions within monomers');
+    ui.tooltip.bind(inputsTabControl.getPane(TAB_DIMERS).header,
+      'Specify symbols for homodimeric and heterodimeric codes');
 
     inputsTabControl.root.style.height = '100%';
     inputsTabControl.root.style.width = '100%';